% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFinaExtract.R
\name{dataFinaExtract}
\alias{dataFinaExtract}
\title{Extraction des données}
\usage{
dataFinaExtract(date_debut, date_fin, NomIndice)
}
\arguments{
\item{date_debut}{La date de début de la période}

\item{date_fin}{La date de fin de la période}

\item{NomIndice}{Le nom de l'indice}
}
\value{
les données boursières entre la date de début et la date de fin
}
\description{
Extraction des données
}
\examples{
DJI <- dataFinaExtract(NomIndice = "^DJI",date_debut = "2021-12-01",date_fin = "2022-01-01")
colnames(DJI) <- c("open","high","low","close","adj_close","volume")

View(DJI)

pour la version data frame:

df_DJI <- data.frame((dataFinaExtract(NomIndice = "^DJI",date_debut = "2021-01-01",date_fin = "2022-01-01")))
colnames(df_DJI) <- c("open","high","low","close","adj_close","volume")

View(df_DJI)
}
