% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-engin.R
\name{feature_engin}
\alias{feature_engin}
\title{Feature Engineering}
\usage{
feature_engin(nomInd, annee)
}
\arguments{
\item{nomInd}{Le nom de l'indice}

\item{annee}{l'année}
}
\value{
Générer les nouvelles variables pour une période donnée
}
\description{
Feature Engineering
}
\examples{
resultat_DJI <- feature_engin(DJI,'2021::')
colnames(resultat_DJI) <- c("Rendement quotidien",
"Rendement hebdomadaire",
"Rendement mensuel",
"Rendement annuel",
"log return quotidien",
"log return hebdo",
"log return mensuel",
"log return annuel",
"Prix cloture moyen de la semaine",
"Prix cloture min de la semaine",
"Prix cloture max de la semaine",
"Prix cloture moyen de la semaine",
"Médiane du prix de cloture de la semaine",
"Volume moyen de la semaine",
"Prix cloture moyen du mois",
"Prix cloture min du mois",
"Prix cloture max du mois",
"Prix cloture moyen du mois",
"Médiane du prix de cloture du mois",
"Volume moyen du mois",
"Prix cloture moyen de l'année",
"Prix cloture min de l'année",
"Prix cloture max de l'année",
"Prix cloture moyen de l'année",
"Médiane du prix de cloture de l'année",
"Volume moyen de l'année")
View(resultat_DJI)
}
