% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele lineaire high-close.R
\name{modele_lineaire_highclose}
\alias{modele_lineaire_highclose}
\title{Modelisation lineaire high close}
\usage{
modele_lineaire_highclose(Ind_data_frame)
}
\arguments{
\item{Ind_data_frame}{Le nom de l'indice version data frame}
}
\value{
le modele lineaire de l'indice qui predit la variable Close dependant de la variable High
}
\description{
Modelisation lineaire high close
}
