% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele lineaire low-close.R
\name{modele_lineaire_lowclose}
\alias{modele_lineaire_lowclose}
\title{Modelisation lineaire low close}
\usage{
modele_lineaire_lowclose(Ind_data_frame)
}
\arguments{
\item{Ind_data_frame}{Le nom de l'indice version data frame}
}
\value{
le modele lineaire de l'indice qui predit la variable Close dependant de la variable Low
}
\description{
Modelisation lineaire low close
}
