% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modele lineaire open-close.R
\name{modele_lineaire_openclose}
\alias{modele_lineaire_openclose}
\title{Modelisation lineaire open close}
\usage{
modele_lineaire_openclose(Ind_data_frame)
}
\arguments{
\item{Ind_data_frame}{Le nom de l'indice version data frame}
}
\value{
le modele lineaire de l'indice qui predit la variable Close dependant de la variable Open
}
\description{
Modelisation lineaire open close
}
